Add automatic casting of int sizes
Neo4j sends us data in the smallest size the value fits in. For example,
if the value is 42, it will serialize it as an 8-bit signed int. The
problem was that if we tried to map it to a different type of int, it
raised an exception. This commit handles the type casting for you.

+++ spec/neo4j/mapping_spec.cr
@@ -7,7 +7,7 @@ module Neo4j
     ::Neo4j.map_node(
       name: String,
       created_at: Time,
-      number: Int32,
+      number: Int64,
       nilable_value: { type: String, nilable: true },
     )
   end
@@ -29,7 +29,7 @@ module Neo4j
         properties: {
           "name" => "Jamie",
           "created_at" => Time.new(2015, 4, 20, 16, 20, 31).to_unix,
-          "number" => 42,
+          "number" => 42_i8,
           "nilable_value" => nil,
         } of String => Type,
       ))
